requirements.txt
flask
flask-sqlalchemy
flask-login
flask-migrate
psycopg2-binary
python-dotenv
reportlab
email-validator